+++ R/reexports.R
@@ -1,0 +1,7 @@
+#' @importFrom generics tidy
+#' @export
+generics::tidy
+
+#' @importFrom ggplot2 autoplot
+#' @export
+ggplot2::autoplot